Median of Two Sorted Arrays - O(log(n)) version
Given two sorted arrays nums1 and nums2 of size m and n respectively, return the median of the two sorted arrays.

The overall run time complexity should be O(log (m+n)).

Since we need to hit O(log(n)) I'm guessing we somehow need to use binary something something, like a binary search.

The key difference over O(n) is that we're doing a binary search on just partial subarrays (partitions) - or finding the partition sizes using binary search to be more precise - and then we're calculating the median based off of just half of the elements.

Finding these boundaries looks like a hack, feels like a hack (or a code-smell at least) but is really not.

The real reason why it *still* looks like a hack is due to the fact that the problem statement *NEVER* mentions "sorted and consecutive", which this solution not only implies but also relies on.

EDIT: I stand corrected, it does not rely on it and works fine for non-consecutive arrays. lol.